requirements.txt
aniso8601==3.0.2
# api==0.0.7
asn1crypto==0.24.0
# certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
click==6.7
config==0.3.9
# cryptography==2.3.1
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dateutils==0.6.6
decorator==4.0.11
dill==0.2.8.2
Django==2.2.7
django-appconf==1.0.2
django-bulk-update==2.2.0
django-constance==2.2.0
django-cors-headers==2.1.0
django-cryptography==0.3
django-extensions==2.0.7
django-filter==2.0.0
django-picklefield==1.0.0
django-redis==4.10.0
django-rq==2.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
# dlib==19.16.99 separately installed from source
docopt==0.6.2
drf-extensions==0.3.1
ExifRead==2.1.2
face-recognition==1.2.2
face-recognition-models==0.3.0
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
geopy==1.11.0
gunicorn==19.8.1
# idna==2.5
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.10
MarkupSafe==1.0
# matplotlib==2.0.2
# mkl-fft==1.0.6
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
networkx==2.2
nltk==3.2.4
nose==1.3.7
olefile==0.44
opencv-python==3.4.3.18
pandas==0.23.4
pexpect==4.6.0
pickleshare==0.7.4
# Pillow==5.1.0
pipreqs==0.4.7
plac==0.9.6
preshed==2.0.1
ptyprocess==0.5.2
pycocotools==2.0.0
psycopg2-binary==2.8.4
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
https://github.com/owncloud/pyocclient/archive/master.zip
pyparsing==2.2.0
python-dateutil==2.7.3
python-memcached==1.58
pytz==2019.3
redis==3.3.11
regex==2018.1.10
# requests==2.18.4
rq==1.1.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.8.1
serpy==0.3.1
simplegeneric==0.8.1
# six==1.10.0
spacy==2.0.18
thinc==6.12.0
toolz==0.9.0
# torch==0.4.1
torchvision==0.2.1
tornado==4.5.1
tqdm==4.14.0
traitlets==4.3.2
ujson==1.35
# urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
yarg==0.1.9
joblib==0.13.2
gevent==1.4.0
numpy==1.17.3
faiss-cpu==1.6.0
jupyter==1.0.0
prompt-toolkit==1.0.15
enzyme==0.4.1
ffmpeg-python==0.2.0
Cython==0.29.14
